Modify app.rb
