+++ data_values.rb
@@ -1,6 +1,7 @@
 module Data
   @books = []
   @peoples = []
+  @rentals = []
 
   def self.books
     @books
@@ -8,5 +9,9 @@ module Data
 
   def self.peoples
     @peoples
+  end
+
+  def  self.rentals
+    @rentals
   end
 end
