+++ rental.rb
@@ -6,9 +6,9 @@ class Rental
     @date = date
 
     @book = book
-    book.rentals << self
+   
 
     @person = person
-    person.rentals << self
+
   end
 end
